refactor : Database 클래스를 mysql 패키지로 이동
