+++ src/main/java/webserver/service/AlreadyLoggedInService.java
@@ -1,10 +1,9 @@
 package webserver.service;
 
-import db.Database;
+import db.mysql.Database;
 import db.UserIdSession;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import webserver.WebServer;
 import webserver.constants.InBody;
 import webserver.httpUtils.Request;
 import webserver.httpUtils.Response;
