+++ src/main/java/webserver/service/LogInService.java
@@ -3,7 +3,7 @@ package webserver.service;
 import customException.cannotLogIn.CannotLogInException;
 import customException.cannotLogIn.NotFoundUserException;
 import customException.cannotLogIn.PasswordMismatchException;
-import db.Database;
+import db.mysql.Database;
 import db.UserIdSession;
 import model.User;
 import org.slf4j.Logger;
